requirements.txt
selenium==4.30.0
webdriver-manager==4.0.2
pytest-html==4.1.1